+++ docs/sphinx/changelog.rst
@@ -16,6 +16,7 @@ Breaking Changes
 New Features
 ============
 - Added calculation of the gradients of the current and previous F1 configurations (:merge:`11`). By `Nathan Miller`_.
+- Added residual class for constructing the residual equations (:merge:`12`). By `Nathan Miller`_.
 
 Internal Changes
 ================
@@ -35,3 +36,4 @@ Internal Changes
 - Added the computation of the gradient of a sub-configuration by all of the configurations (:merge:`9`). By `Nathan Miller`_.
 - Added gradients for the preceding and following sub-configurations for the current and previous configurations (:merge:`9`). By `Nathan Miller`_.
 - Required >= version 0.5.3 of vector_tools (:merge:`10`). By `Nathan Miller`_.
+- Added the construction residual, Jacobian, and other values (:merge:`13`). By `Nathan Miller`_.